+++ src/ui/SftpBrowserTab.h
@@ -6,10 +6,13 @@
 #include <QString>
 #include <QWidget>
 
+class QFileSystemModel;
 class QLabel;
 class QLineEdit;
 class QPushButton;
+class QModelIndex;
 class QTableWidget;
+class QTreeView;
 
 class SftpBrowserTab : public QWidget
 {
@@ -29,13 +32,21 @@ public:
 
 private:
     void setupUi();
-    void refreshDirectory();
-    void openPathFromEditor();
-    void goUp();
-    void handleCellDoubleClicked(int row, int column);
-    void populateTable(const QList<SftpRemoteEntry> &entries);
-    void setBusy(bool busy);
-    QString normalizedPath(const QString &path) const;
+
+    void openLocalPathFromEditor();
+    void goLocalUp();
+    void refreshLocalDirectory();
+    void handleLocalDoubleClicked(const QModelIndex &index);
+    void setLocalPath(const QString &path);
+    QString normalizedLocalPath(const QString &path) const;
+
+    void refreshRemoteDirectory();
+    void openRemotePathFromEditor();
+    void goRemoteUp();
+    void handleRemoteCellDoubleClicked(int row, int column);
+    void populateRemoteTable(const QList<SftpRemoteEntry> &entries);
+    void setRemoteBusy(bool busy);
+    QString normalizedRemotePath(const QString &path) const;
     QString joinedRemotePath(const QString &basePath, const QString &entryName) const;
     QString parentRemotePath(const QString &path) const;
     QString statusPrefix() const;
@@ -47,12 +58,22 @@ private:
     SshAuthMethod m_authMethod = SshAuthMethod::Password;
     QString m_secretValue;
     SshHostKeyExpectation m_hostKeyExpectation;
-    QString m_currentPath = QStringLiteral(".");
 
-    QLineEdit *m_pathEdit = nullptr;
-    QPushButton *m_goButton = nullptr;
-    QPushButton *m_upButton = nullptr;
-    QPushButton *m_refreshButton = nullptr;
-    QLabel *m_statusLabel = nullptr;
-    QTableWidget *m_table = nullptr;
+    QString m_currentLocalPath;
+    QString m_currentRemotePath = QStringLiteral(".");
+
+    QLineEdit *m_localPathEdit = nullptr;
+    QPushButton *m_localGoButton = nullptr;
+    QPushButton *m_localUpButton = nullptr;
+    QPushButton *m_localRefreshButton = nullptr;
+    QLabel *m_localStatusLabel = nullptr;
+    QFileSystemModel *m_localModel = nullptr;
+    QTreeView *m_localTree = nullptr;
+
+    QLineEdit *m_remotePathEdit = nullptr;
+    QPushButton *m_remoteGoButton = nullptr;
+    QPushButton *m_remoteUpButton = nullptr;
+    QPushButton *m_remoteRefreshButton = nullptr;
+    QLabel *m_remoteStatusLabel = nullptr;
+    QTableWidget *m_remoteTable = nullptr;
 };